De-duplicate code around getting bt fields
